Исправлены баги: убран поиск + исправен баг с получением друзей (в постгре не было приведения типа)

+++ internal/repositories/friends_repo.go
@@ -85,7 +85,7 @@ func (r *UserRepository) GetFriends(userID int) ([]models.Friend, error) {
 	query := `
 		SELECT 
 			f.id,
-			$1 as user_id,
+			$1::int as user_id,
 			CASE 
 				WHEN f.user_id = $1 THEN f.friend_id 
 				ELSE f.user_id
